src/backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
deta==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
isodate==0.6.1
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
python-youtube==0.9.4
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
yarl==1.9.4
